% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{default_workflow_map}
\alias{default_workflow_map}
\title{Default workflow step map}
\usage{
default_workflow_map()
}
\value{
Named character vector.
}
\description{
Returns the default named map of workflow steps to project script paths.
}
