% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{find_project_root}
\alias{find_project_root}
\title{Find a project root directory}
\usage{
find_project_root(start = getwd(), max_up = 10L)
}
\arguments{
\item{start}{Starting directory.}

\item{max_up}{Maximum number of parent traversals.}
}
\value{
Absolute path to the detected project root.
}
\description{
Searches upward from \code{start} for a directory containing a Quarto config
and a \code{data/public} tree.
}
