% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{summarize_output_audit}
\alias{summarize_output_audit}
\title{Summarize an output audit}
\usage{
summarize_output_audit(audit_tbl)
}
\arguments{
\item{audit_tbl}{Result from \code{\link[=audit_public_outputs]{audit_public_outputs()}}.}
}
\value{
Named list with high-level diagnostics.
}
\description{
Summarize an output audit
}
